requirements.txt
-i https://pypi.org/simple
beautifulsoup4==4.11.2
cachetools==5.3.0 ; python_version ~= '3.7'
cchardet==2.1.7
certifi==2022.12.7 ; python_version >= '3.6'
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
google-api-core==2.11.0 ; python_version >= '3.7'
google-api-python-client==2.77.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.58.0 ; python_version >= '3.7'
halo==0.0.31
httplib2==0.21.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
idna==3.4 ; python_version >= '3.5'
log-symbols==0.0.14
lxml==4.9.2
oauthlib==3.2.2 ; python_version >= '3.6'
protobuf==4.21.12 ; python_version >= '3.7'
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9 ; python_version >= '3.1'
python-dateutil==2.8.2
requests==2.28.2
requests-oauthlib==1.3.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
rsa==4.9 ; python_version >= '3.6'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
soupsieve==2.3.2.post1 ; python_version >= '3.6'
spinners==0.0.24
termcolor==2.2.0 ; python_version >= '3.7'
uritemplate==4.1.1 ; python_version >= '3.6'
urllib3==1.26.14 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
